[WIP] [Update] Support for Tundra Tracker quirks
WORK IN PROGRESS - DO NOT MERGE - INCOMPLETE

This does NOT fix the "Identify" button!

Tentative intensity multiplier value: 9.0
"Linear" is strong, "Throb" is a bit weak, but that may change when
the means of triggering haptic pulses is better tuned to support
Tundra Trackers.

This might break Vive Tracker support, or at the very least needlessly
use more CPU cycles with the tighter FeedbackThread loop.

This might need replaced with support for the IVRInput system in place
of the deprecated triggerHapticPulse() function.

This is just the bare minimum to get Tundra Trackers working with the
bridge app during normal use (not the Identify button).

Haptic Tundrakes, anyone? ...no?