[Update] Support Tundra Tracker with timing quirks
Support the Tundra Tracker using microseconds instead of milliseconds
for the legacy, deprecated "triggerHapticPulse()" method.  This also
comes with an upper limit of around 4000 µs (4 ms), so the update loop
must run more frequently to stay within that maximum duration for a
single pulse.

Set Tundra Trackers back to a multiplier of 1.0, matching the behavior
of the Vive Trackers.

This might need replaced with support for the IVRInput system in place
of the deprecated triggerHapticPulse() function.

Warn if attempting to set a strength exceeding 100% (1.0).  It should
not happen as that should have no effect - the pulse duration lasts
longer than the delay before the next pulse.  However, multipliers
resulting in excess of 100% offer a way to adust for odd trackers
without needing to rebuild the released app on Windows.

Haptic Tundrakes, anyone? ...no?

Tested setup:
* 4x pancake vibration motors
   Thinner than LRAs shipped with Tundra IO Boards, fits in stock base

* 4x Tundra Trackers
   Left/right foot, waist, chest

* Default multiplier, default strengths
   Also tested Velocity strength set to 60% - 100%

* Proximity and Velocity set to "Linear" pattern
   "Throb" works too, but feels a bit weak