[Add] OSCQuery for VRChat, auto-detect port
Add OSCQuery support VRChat, including auto-detecting an available
port.  This avoids the need for OSC router applications.

Adds dependency on CyberKitsune's alpha "tinyoscquery"

Rework start/stop to reduce freezing on main thread.

NOTE: mDNS/zeroconf names must be unique.  Ideally, this would be
entirely local.  Unfortunately, attempts to get `zeroconf` to only
advertise locally have been met with pain and suffering.

mDNS is not even consistent with service name conflict detection,
sometimes claiming a name is available when it's in use.  Even on the
same machine!

To try to minimize failure, always append a random 3-digit suffix to
the service name.  It's not pretty, but it works.  And re-randomize up
to five times if needed.

Hat tip to qDot for reassuring me that I was not in fact losing it,
mDNS is simply cursed.

See https://github.com/cyberkitsune/tinyoscquery