[Add] Clear stuck haptics if input stops for X sec
Add new option to clear haptics output when input stops for a period
of time, e.g. if Resonite/VRChat crashes, switching avatars, or
something just breaks.

Of note, VRChat stops sending updates when a value is peaked at 1.0,
so that could be given a longer delay to help avoid wrongly clearing the
value.  This hasn't seemed to matter much in practice though and may add
confusion.